remove the keyGen.go file and move key generation interface to key.go file and move its implementations for rsa, ecdsa to rsaKey.go, ecdsaKey.go files.
